[Refactor] Home View UI 파일 이름 수정
