+++ Harmony/Harmony/Models/GroupModel.swift
@@ -13,13 +13,6 @@ struct User: Identifiable {
     var profile: String?
 }
 
-//struct FamilyMember: Identifiable {
-//    let id: String
-//    var user: User
-//    let permissionId: String
-//    var alias: String?
-//}
-
 struct FamilyMember: Identifiable {
     let id: String
     var user: User
@@ -27,4 +20,18 @@ struct FamilyMember: Identifiable {
     var alias: String?
     var groupId: Int
     var deviceToken: String?
+}
+
+enum UserRole: String {
+    case member = "M"
+    case vip = "V"
+    
+    var name: String {
+        switch self {
+        case .member:
+            return "Member"
+        case .vip:
+            return "VIP"
+        }
+    }
 }
